+++ atrium-api/src/app/bsky/feed/get_reposted_by.rs
@@ -5,7 +5,7 @@
 pub struct Parameters {
     ///If supplied, filters to reposts of specific version (by CID) of the post record.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cid: Option<String>,
+    pub cid: Option<crate::types::string::Cid>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cursor: Option<String>,
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -17,7 +17,7 @@ pub struct Parameters {
 #[serde(rename_all = "camelCase")]
 pub struct Output {
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub cid: Option<String>,
+    pub cid: Option<crate::types::string::Cid>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cursor: Option<String>,
     pub reposted_by: Vec<crate::app::bsky::actor::defs::ProfileView>,