+++ atrium-api/src/com/atproto/sync/get_blocks.rs
@@ -3,7 +3,7 @@
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
 pub struct Parameters {
-    pub cids: Vec<String>,
+    pub cids: Vec<crate::types::string::Cid>,
     ///The DID of the repo.
     pub did: crate::types::string::Did,
 }