+++ atrium-api/src/com/atproto/sync/get_record.rs
@@ -6,7 +6,7 @@ pub struct Parameters {
     pub collection: crate::types::string::Nsid,
     ///An optional past commit CID.
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub commit: Option<String>,
+    pub commit: Option<crate::types::string::Cid>,
     ///The DID of the repo.
     pub did: crate::types::string::Did,
     ///Record Key