+++ atrium-api/src/com/atproto/sync/list_blobs.rs
@@ -16,7 +16,7 @@ pub struct Parameters {
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
 pub struct Output {
-    pub cids: Vec<String>,
+    pub cids: Vec<crate::types::string::Cid>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub cursor: Option<String>,
 }