+++ lexicon/atrium-codegen/src/token_stream.rs
@@ -491,6 +491,7 @@ fn string_type(string: &LexString) -> Result<(TokenStream, TokenStream)> {
     // TODO: enum?
     let typ = match string.format {
         Some(LexStringFormat::AtIdentifier) => quote!(crate::types::string::AtIdentifier),
+        Some(LexStringFormat::Cid) => quote!(crate::types::string::Cid),
         Some(LexStringFormat::Datetime) => quote!(crate::types::string::Datetime),
         Some(LexStringFormat::Did) => quote!(crate::types::string::Did),
         Some(LexStringFormat::Handle) => quote!(crate::types::string::Handle),